backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
sentence-transformers==2.2.2
numpy==1.26.2
scikit-learn==1.3.2
pandas==2.1.3
python-multipart==0.0.6
requests==2.31.0
aiohttp==3.9.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
faiss-cpu==1.7.4
openai==1.3.0
matplotlib==3.8.2
seaborn==0.13.0
schedule==1.2.1
spacy==3.7.2
sqlalchemy==2.0.23
redis==5.0.1
psycopg2-binary==2.9.9 